write_ops and root dot dir entry writing

+++ hw2/common/superblock.h
@@ -28,7 +28,7 @@ int  write_superblock(const struct Superblock *superblock, FILE *file);
 int  read_superblock (struct Superblock *superblock, FILE *file);
 void free_superblock (const struct Superblock *superblock);
 
-int set_block_use(struct Superblock *superblock, const size_t block_id, const int is_used);
-int set_inode_use(struct Superblock *superblock, const size_t inode_id, const int is_used);
+int set_block_use(struct Superblock *superblock, const uint32_t block_id, const int is_used);
+int set_inode_use(struct Superblock *superblock, const uint32_t inode_id, const int is_used);
 
 #endif
